requirements.txt
requests==2.32.3
python-dotenv==1.0.1
pycryptodome==3.20.0
pyyaml==6.0.2
pyinstaller>=6.15.0